bug fixed add reqwards and payment using rewards api

+++ spring-payments/src/main/java/com/example/springpayments/Customer.java
@@ -38,7 +38,7 @@ public class Customer {
     private String zip;
     private String phoneNumber;
     private String email;
-    private int rewards;
+    private double rewards;
 
 
     @OneToMany(mappedBy="customer", fetch = FetchType.EAGER, orphanRemoval = true)
